# openapi/windows-ai.yaml
openapi: 3.1.0
info:
  title: Windows AI Actions API
  version: 0.2.0
paths:
  /api/actions/execute:
    post:
      summary: Execute a system action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../codex/SCHEMAS/actions.request.schema.json
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: ../codex/SCHEMAS/actions.response.schema.json
...